+++ packages/tests-react-native/entry.js
@@ -35,3 +35,6 @@ if (global.process.stdout === undefined) {
         isTTY: false,
     }
 }
+
+export default function() {
+}